Great leaders are willing to sacrifice the numbers to save the people. Poor leaders sacrifice the people to save the numbers updated final

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 int _atoi(char *s)
 {
 	int sign = 1;
-	int result = 0;
+	long int result = 0;
 
 	while (*s && (*s < '0' || *s > '9'))
 	{
